Optimized the 'process_image.cpp' code.
Initially, I was first checking that in the image, whether there is
a white ball or not. Then I was deciding it is in which part. But then
I found that it's unnecessary. So, now I'm checking ball is in which
part. If it's in none of the parts of an image which means that it is
not in the image.